Id由Long修改为Int（用不到Long）

+++ BellaVoice/app/src/main/java/com/example/bellavoice/model/SongBean.kt
@@ -1,7 +1,7 @@
 package com.example.bellavoice.model
 
 data class SongBean @JvmOverloads constructor(
-    val id: Long = 0L,
+    val id: Int = 0,
     val uri: String = "",
     val song: String = "", //歌名
     val singer: String = "", //歌手名称
